requirements.txt
django==5.0.*
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0
django_tinymce==4.1.0